refactor: breakdown the header component into smaller component
