+++ src/App.jsx
@@ -1,4 +1,4 @@
-import Header from '../src/components/Header'
+import Header from './components/Header/Header'
 
 import "bootstrap/dist/css/bootstrap.min.css";
 import "bootstrap/dist/js/bootstrap.bundle.min";
@@ -9,7 +9,7 @@ function App() {
   return (
     <>
     <Header />
-      <h1>hello world</h1>
+      {/* <h1>hello world</h1> */}
     </>
   )
 }
