+++ src/components/Header/Header.css
@@ -3,14 +3,7 @@ header {
     padding: 10px 0;
 }
 
-.Logo {
-    height: 50px;
-}
 
-.logo-text {
-    font-size: 1.2rem;
-    color: #333;
-}
 
 .nav-link {
     margin-right: 20px;
