+++ src/components/Header/components/HeaderLogo.css
@@ -1,0 +1,8 @@
+.Logo {
+    height: 50px;
+}
+
+.logo-text {
+    font-size: 1.2rem;
+    color: #333;
+}
